Implement serialization
When doing this I realized that the frame data in the motion section is actually supposed to be layed out so that all of the floats corresponding to a particular frame are on the same line. This means the parser is currently discarding this semantic information. For now I've reconstructed it in the serializer (and added the helpful Joint::total_channels fn to calculate how many channels exist for a given joint and its subtree) but I think the proper approach should be to keep the frame values separated in the Bvh structure.